refactor: 폴더 생성 (#122)
